Add DRS generate UI, improve API/logging & models
Add new DRS generate screen and related controller changes, rename/move drs_generate files for cleaner module layout. Improve WebService: initialize logger, set Dio LogInterceptor and move initialization call to main. Enhance DRSGenerateController with form validation, manual vendor/vehicle support, cnote add flow, submit flow with loading state and user feedback dialogs, and more detailed logging. Change package count fields from int to double in available/prepare DRS models to handle fractional values. Update Get routes/imports, add devtools config, and enhance TmsButton to support loading state, styling options and disabled behavior.

+++ lib/model/drs_model/drs_generate/available_docket_model.dart
@@ -42,8 +42,8 @@ class AvailableDocketItem {
   String desT_CD;
   String curr_Loc;
   int pendPkgQty;
-  int arrPkgQty;
-  int pkgsno;
+  double arrPkgQty;
+  double pkgsno;
   String payBas;
   String paybaS_Str;
   String cdeldt;
@@ -109,8 +109,8 @@ class AvailableDocketItem {
     desT_CD: json["desT_CD"] ?? "",
     curr_Loc: json["curr_Loc"] ?? "",
     pendPkgQty: json["pendPkgQty"] ?? 0,
-    arrPkgQty: json["arrPkgQty"] ?? 0,
-    pkgsno: json["pkgsno"] ?? 0,
+    arrPkgQty: (json["arrPkgQty"] ?? 0).toDouble(),
+    pkgsno: (json["pkgsno"] ?? 0).toDouble(),
     payBas: json["payBas"] ?? "",
     paybaS_Str: json["paybaS_Str"] ?? "",
     cdeldt: json["cdeldt"] ?? "",
