+++ lib/model/drs_model/drs_generate/prepare_drs_model.dart
@@ -120,8 +120,8 @@ class DrsGenerateItem {
   String dockno;
   String docksf;
   String orgncd;
-  int pkgsno;
-  int arrPkgQty;
+  double pkgsno;
+  double arrPkgQty;
   int pendPkgQty;
   String payBas;
   double actuwt;
